Pare down git alias command, expand on its TODOs
As-is, the `column -t` invocation works very poorly and is effectively
useless as it can't communicate the information it ought to. For the
time being (until the next Debian stable coreutils update) the format
`alias = 'cmd'` is more readable and useful. This commit also expands a
good bit on the ideal command chain of the post-coreutils update alias
and the possible issues left to solve after that implementation is done.