Add nohup convenience alias
This alias removes any output which would be generated. By default,
`nohup` generates a `nohup.out` file in the cwd or home directory with
the output from the executed program. Most of the time I don't need to
monitor the program's output, and when I do, I can easily run `\nohup
<program>` to circumvent this alias.