Explicitly include user gitconfig
This bypasses an annoying issue when using both ~/.gitconfig and
~/.user.gitconfig, which is that there is no guarantee of the order
these configs are loaded. This means that configs loaded in
~/.user.gitconfig (which in this case is intended to override
~/.gitconfig) may not actually override those found in ~/.gitconfig. By
using `include.path`, they are loaded a second time, but it guarantees
they are loaded after all the configs in ~/.gitconfig, ensuring they are
used in cases where the key is present in both files.