Fix bashrc .complete completion behavior
This commit updates the completion function for the .functions (`..`,
`.2`, etc.) to behave as much like tab completion for `cd` as possible.
This means it has better listing for possible results when multiple
options are available, properly handles completed and nested directories
(!!) and is much more thoroughly documented so it's not such a dense
cluster of bash magic.