Add libinput-gestures config
This commit also adds an old script used for facilitating browser
navigation backwards with a 4-finger swipe, as Firefox previously was
not handling overscroll navigation correctly. It is no longer in use as
the fix in the comments of these files made precise scrolling work,
which in turn made overscroll navigation work as well.