Reduce RuneLite rebuild script max age to 18 hours
This will mean that a client built one morning will almost always yield
another build the next day (assuming there are changes for it to be
necessary) even if it goes from being built late one morning to early
the next morning. Despite that, 18 hours is still long enough that a
client build will be used for roughly a full day.