Fail music transfer when local file lookup fails
Prior to this change, a transfer file with missing directories and files
would silently go on without any kind of notice that these files were
not included. To ensure the file is up-to-date with any name changes or
misspellings of source files, these should cause an error and halt prior
to transfer.