Don't source .profile from .bash_profile
bash has [a specific order in which it tries to source a profile/login
file][1], so given that this repository provides a .bash_profile file,
it will not attempt to, and in fact should not, source .profile. This
was causing some bad behavior on login to some machines which had a
system default ~/.profile provided which re-sourced .bashrc.

[1]: https://www.gnu.org/savannah-checkouts/gnu/bash/manual/bash.html#Invoked-as-an-interactive-login-shell_002c-or-with-_002d_002dlogin